backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4
openai-whisper==20231117
moviepy==1.0.3
ffmpeg-python==0.2.0
celery==5.3.4
redis==5.0.1
stripe==7.8.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
Pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchaudio==2.1.1
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
websockets==12.0
jinja2==3.1.2
python-slugify==8.0.1
groq==0.4.1
